Corrected the push/pull margin calculations.
The previous formula for calculating push/pull margins was incorrect.
The correct behavior is for the push/pull margin of $x to be equivalent
to the width of column($x) plus half of the gutter.

The formula could be simplified further, but I feel that writing it this
way emphesises the relationship to the column width formula.